Finish the API and test the route with postman

+++ app/Http/Controllers/API/AddressController.php
@@ -44,11 +44,20 @@ class AddressController extends Controller
     public function show(Address $address)
     {
         try {
-            return response()->json([
+            if(Auth::user()->id === $address->user_id || Auth::user()->role_id === 2) {
+                return response()->json([
                 'code' => 200,
                 'status' => 'success',
                 'data' => $address,
             ]);
+            } else {
+                return response()->json([
+                    'code' => 401,
+                    'status' => 'error',
+                    'message'=> 'Pas autorisé'
+                    ]);
+            }
+
         } catch (Exception $e) {
             return response()->json([
                  'code' => 404,
@@ -65,7 +74,8 @@ class AddressController extends Controller
     public function update(Request $request, Address $address)
     {
         try {
-            $request->validate([
+            if(Auth::user()->id === $address->user_id) {
+                $request->validate([
                 'address_name' => 'required|string|min:3|max:50',
                 'address' => 'required|string|min:3|max:100',
                 'postal_code' => 'required|string|size:5',
@@ -73,13 +83,20 @@ class AddressController extends Controller
                 'country' => 'required|string|min:3|max:50',
             ]);
             $address->update($request->all());
-
             return response()->json([
                 'code' => 201,
                 'status' => 'success',
                 'data' => $address,
                 'message'=> 'Mise à jour de l\'adresse avec succès'
             ]);
+
+            } else {
+                return response()->json([
+                'code' => 401,
+                'status' => 'error',
+                'message'=> 'Pas autorisé'
+                ]);
+                }
             } catch (Exception $e) {
                 return response()->json([
                     'code' => 404,
@@ -96,12 +113,20 @@ class AddressController extends Controller
     public function destroy(Address $address)
     {
         try {
+            if(Auth::user()->id === $address->user_id) {
             $address->delete();
             return response()->json([
                 'code' => 201,
                 'status' => 'success',
                 'message' => 'Suppression réussite'
             ]);
+            } else {
+                return response()->json([
+                    'code' => 401,
+                    'status' => 'error',
+                    'message'=> 'Pas autorisé'
+                    ]);
+            }
         } catch (Exception $e) {
             return response()->json([
                  'code' => 404,
