+++ app/Http/Controllers/API/CategoryController.php
@@ -34,16 +34,24 @@ class CategoryController extends Controller
     public function store(Request $request)
     {
         try {
-             $request->validate([
-            'category_name' => 'required|min:3|max:50|unique:categories'
-        ]);
-            $category = Category::create($request->all());
-            return response()->json([
-                'code' => 201,
-                'status' => 'success',
-                'data' => $category,
-                'message'=> 'Ajout de la catégorie avec succès'
-            ]);
+            if(Auth::user()->role_id === 2) {
+                $request->validate([
+                    'category_name' => 'required|min:3|max:50|unique:categories'
+                    ]);
+                $category = Category::create($request->all());
+                return response()->json([
+                    'code' => 201,
+                    'status' => 'success',
+                    'data' => $category,
+                    'message'=> 'Ajout de la catégorie avec succès'
+                ]);
+            } else {
+                return response()->json([
+                    'code' => 401,
+                    'status' => 'error',
+                    'message' => 'Pas autorisé',
+                ]);
+            }
         } catch (Exception $e) {
             return response()->json([
                 'code' => 404,
@@ -81,17 +89,25 @@ class CategoryController extends Controller
     public function update(Request $request, Category $category)
     {
         try {
-            $request->validate([
-                'category_name' => 'required|min:3|max:50|unique:categories'
-            ]);
-            $category->update($request->all());
+            if(Auth::user()->role_id === 2) {
+                $request->validate([
+                    'category_name' => 'required|min:3|max:50|unique:categories'
+                ]);
+                $category->update($request->all());
 
-            return response()->json([
-                'code' => 201,
-                'status' => 'success',
-                'data' => $category,
-                'message'=> 'Mise à jour de la catégorie avec succès'
-            ]);
+                return response()->json([
+                    'code' => 201,
+                    'status' => 'success',
+                    'data' => $category,
+                    'message'=> 'Mise à jour de la catégorie avec succès'
+                ]);
+            } else {
+                return response()->json([
+                    'code' => 401,
+                    'status' => 'error',
+                    'message' => 'Pas autorisé',
+                ]);
+            }
         } catch (Exception $e) {
             return response()->json([
                 'code' => 404,
@@ -108,12 +124,20 @@ class CategoryController extends Controller
     public function destroy(Category $category)
     {
         try {
-            $category->delete();
-            return response()->json([
-                'code' => 201,
-                'status' => 'success',
-                'message' => 'Suppression réussite'
-            ]);
+            if(Auth::user()->role_id === 2) {
+                $category->delete();
+                return response()->json([
+                    'code' => 201,
+                    'status' => 'success',
+                    'message' => 'Suppression réussite'
+                ]);
+            } else {
+                return response()->json([
+                    'code' => 401,
+                    'status' => 'error',
+                    'message' => 'Pas autorisé',
+                ]);
+            }
         } catch (Exception $e) {
             return response()->json([
                  'code' => 404,
